Total births per municipality in Switzerland between 1969 and 2012

Each row contains the number of births for a municipality and a year. Not all municipalities are present for each year due to changes in the definition of the municipalities.

SwissBirths

Format

A data frame with 105058 observations of 4 variables:

 $ Year            : Factor w/ 44 levels "1969","1970",..: 1 1 1 1 1 1 1 1 1 1 ...
 $ MunicipalityID  : Factor w/ 2485 levels "1","2","3","4",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ MunicipalityName: Factor w/ 2485 levels "Aeugst am Albis",..: 1 2 3 4 5 6 7 8 9 10 ...
 $ Births          : num  13 174 22 36 20 13 17 4 32 48 ...

Source

STAT-TAB, section 01.2 "Bevölkerung/Bevölkerungsstand und -bewegung", item "Lebendgeburten nach institutionellen Gliederungen, Geschlecht und Staatsangehörigkeit des Kindes und nach Altersklasse der Mutter".

STAT-TAB URL: http://www.pxweb.bfs.admin.ch/Database/German_01%20-%20Bev%C3%B6lkerung/01.2%20-%20Bev%C3%B6lkerungsstand%20und%20-bewegung/01.2%20-%20Bev%C3%B6lkerungsstand%20und%20-bewegung.asp?lang=1&prod=01&secprod=2&openChild=true.

File link: http://www.pxweb.bfs.admin.ch/Database/German_01%20-%20Bev%C3%B6lkerung/01.2%20-%20Bev%C3%B6lkerungsstand%20und%20-bewegung/px-d-01-2D02.px.

Author

Swiss Federal Statistical Office