Create a list of municipality mutations

Municipality mutation lists are required to compute matching tables for municipalities between different points in time.

swcGetMutations(swc = NULL)

Arguments

swc
Ignored.

Value

A data frame that represents mutations.

Details

The municipality dataset of the Swiss historicized communes data is a list of snapshots of municipality states. Each state is valid during a certain period of time, adjoining states are linked by admission and abolition numbers: The abolition number of a former state is the same as the admission number of a subsequent state. The states can be thought of as nodes in a graph, where the edges are mutations -- transformations from one state to the next.

This function performs a self-merge on the municipality data: Each abolition number is matched with its corresponding admission number (if available). If no corresponding admission or abolition number is found, the record is included with NAs instead of matched values. Records without admission or abolition number are excluded. The result is a list of mutations, i.e., a list of edges in the graph of municipality state snapshots.

Examples

head(swcGetMutations(), 20)

#>    mMutationNumber mHistId.x mId.x mShortName.x mAbolitionMode mAbolitionDate
#> 1             1000        NA    NA         <NA>           <NA>           <NA>
#> 2             1000        NA    NA         <NA>           <NA>           <NA>
#> 3             1000        NA    NA         <NA>           <NA>           <NA>
#> 4             1000        NA    NA         <NA>           <NA>           <NA>
#> 5             1000        NA    NA         <NA>           <NA>           <NA>
#> 6             1000        NA    NA         <NA>           <NA>           <NA>
#> 7             1000        NA    NA         <NA>           <NA>           <NA>
#> 8             1000        NA    NA         <NA>           <NA>           <NA>
#> 9             1000        NA    NA         <NA>           <NA>           <NA>
#> 10            1000        NA    NA         <NA>           <NA>           <NA>
#> 11            1000        NA    NA         <NA>           <NA>           <NA>
#> 12            1000        NA    NA         <NA>           <NA>           <NA>
#> 13            1000        NA    NA         <NA>           <NA>           <NA>
#> 14            1000        NA    NA         <NA>           <NA>           <NA>
#> 15            1000        NA    NA         <NA>           <NA>           <NA>
#> 16            1000        NA    NA         <NA>           <NA>           <NA>
#> 17            1000        NA    NA         <NA>           <NA>           <NA>
#> 18            1000        NA    NA         <NA>           <NA>           <NA>
#> 19            1000        NA    NA         <NA>           <NA>           <NA>
#> 20            1000        NA    NA         <NA>           <NA>           <NA>
#>    mDateOfChange.x mHistId.y mId.y       mShortName.y          mAdmissionMode
#> 1             <NA>     11348     1             Aeugst First-time registration
#> 2             <NA>     11742     2 Affoltern am Albis First-time registration
#> 3             <NA>     11801     3         Bonstetten First-time registration
#> 4             <NA>     11992     4    Hausen am Albis First-time registration
#> 5             <NA>     12249     5           Hedingen First-time registration
#> 6             <NA>     12671     6    Kappel am Albis First-time registration
#> 7             <NA>     12674     7             Knonau First-time registration
#> 8             <NA>     12738     8        Maschwanden First-time registration
#> 9             <NA>     12433     9     Mettmenstetten First-time registration
#> 10            <NA>     12497    10           Obfelden First-time registration
#> 11            <NA>     13124    11          Ottenbach First-time registration
#> 12            <NA>     12891    12         Rifferswil First-time registration
#> 13            <NA>     10575    13          Stallikon First-time registration
#> 14            <NA>     11349    14           Wettswil First-time registration
#> 15            <NA>     11735    21            Adlikon First-time registration
#> 16            <NA>     11862    22        Benken (ZH) First-time registration
#> 17            <NA>     11855    23     Berg am Irchel First-time registration
#> 18            <NA>     11480    24     Buch am Irchel First-time registration
#> 19            <NA>     11573    25            Dachsen First-time registration
#> 20            <NA>     12182    26               Dorf First-time registration
#>    mAdmissionDate mDateOfChange.y mMutationDate     mMutationId
#> 1      1960-01-01      1976-11-14    1960-01-01 1960-01-01.1000
#> 2      1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 3      1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 4      1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 5      1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 6      1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 7      1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 8      1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 9      1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 10     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 11     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 12     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 13     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 14     1960-01-01      1976-11-14    1960-01-01 1960-01-01.1000
#> 15     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 16     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 17     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 18     1960-01-01      2012-12-31    1960-01-01 1960-01-01.1000
#> 19     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000
#> 20     1960-01-01      1960-01-01    1960-01-01 1960-01-01.1000

head(subset(swcGetMutations(), !is.na(mHistId.x)), 20)

#>      mMutationNumber mHistId.x mId.x          mShortName.x
#> 3144            1001     11248  5512               Bercher
#> 3145            1002     11245  2451            Ichertswil
#> 3146            1002     11244  2455            Lüterkofen
#> 3147            1003     11242  5625      Bussy-sur-Morges
#> 3148            1003     11243  5626 Chardonney-sur-Morges
#> 3149            1004     11240  4021                 Baden
#> 3150            1004     11227  4025               Dättwil
#> 3151            1005     11238  5886    Montreux-Châtelard
#> 3152            1005     11253  5887     Montreux-Planches
#> 3153            1006     11237  3972    Seewis im Prätigau
#> 3154            1007     11236  3881   Conters im Prätigau
#> 3155            1008     11235  2295         Grossbösingen
#> 3156            1009     11234  3982       Disentis/Mustèr
#> 3157            1010     11231  3597                Peiden
#> 3158            1010     11230  3602       Uors (Lumnezia)
#> 3159            1011     11233  2935            Herblingen
#> 3160            1011     11232  2939          Schaffhausen
#> 3161            1012     11229  3404                 Henau
#> 3162            1013     11239  4756     Schönholzerswilen
#> 3163            1013     11266  4757                  Toos
#>                 mAbolitionMode mAbolitionDate mDateOfChange.x mHistId.y mId.y
#> 3144 Reassignment to other d/c     1960-12-31      1960-12-31     13186  5512
#> 3145                 Abolition     1960-12-31      1960-12-31     13187  2455
#> 3146                 Abolition     1960-12-31      1960-12-31     13187  2455
#> 3147                 Abolition     1961-03-31      1961-03-31     13188  5625
#> 3148                 Abolition     1961-03-31      1961-03-31     13188  5625
#> 3149            Change of area     1961-12-31      1961-12-31     13189  4021
#> 3150                 Abolition     1961-12-31      1961-12-31     13189  4021
#> 3151                 Abolition     1961-12-31      1961-12-31     13190  5886
#> 3152                 Abolition     1961-12-31      1961-12-31     13190  5886
#> 3153        Change of name (m)     1961-12-31      1961-12-31     13191  3972
#> 3154        Change of name (m)     1961-12-31      1961-12-31     13192  3881
#> 3155        Change of name (m)     1961-12-31      1961-12-31     13193  2295
#> 3156        Change of name (m)     1962-12-31      1962-12-31     13194  3982
#> 3157                 Abolition     1963-05-23      1963-05-23     13195  3602
#> 3158                 Abolition     1963-05-23      1963-05-23     13195  3602
#> 3159                 Abolition     1963-12-31      1963-12-31     13196  2939
#> 3160            Change of area     1963-12-31      1963-12-31     13196  2939
#> 3161        Change of name (m)     1964-03-09      1964-03-09     13197  3408
#> 3162            Change of area     1964-05-31      1964-05-31     13198  4756
#> 3163                 Abolition     1964-05-31      1964-05-31     13198  4756
#>               mShortName.y            mAdmissionMode mAdmissionDate
#> 3144               Bercher Reassignment to other d/c     1961-01-01
#> 3145 Lüterkofen-Ichertswil                  Creation     1961-01-01
#> 3146 Lüterkofen-Ichertswil                  Creation     1961-01-01
#> 3147      Bussy-Chardonney                  Creation     1961-04-01
#> 3148      Bussy-Chardonney                  Creation     1961-04-01
#> 3149                 Baden            Change of area     1962-01-01
#> 3150                 Baden            Change of area     1962-01-01
#> 3151              Montreux                  Creation     1962-01-01
#> 3152              Montreux                  Creation     1962-01-01
#> 3153   Seewis im Prättigau        Change of name (m)     1962-01-01
#> 3154  Conters im Prättigau        Change of name (m)     1962-01-01
#> 3155              Bösingen        Change of name (m)     1962-01-01
#> 3156       Disentis/Mustér        Change of name (m)     1963-01-01
#> 3157           Uors-Peiden                  Creation     1963-05-24
#> 3158           Uors-Peiden                  Creation     1963-05-24
#> 3159          Schaffhausen            Change of area     1964-01-01
#> 3160          Schaffhausen            Change of area     1964-01-01
#> 3161                 Uzwil        Change of name (m)     1964-03-10
#> 3162     Schönholzerswilen            Change of area     1964-06-01
#> 3163     Schönholzerswilen            Change of area     1964-06-01
#>      mDateOfChange.y mMutationDate     mMutationId
#> 3144      2006-08-31    1961-01-01 1961-01-01.1001
#> 3145      1961-01-01    1961-01-01 1961-01-01.1002
#> 3146      1961-01-01    1961-01-01 1961-01-01.1002
#> 3147      2006-08-31    1961-04-01 1961-04-01.1003
#> 3148      2006-08-31    1961-04-01 1961-04-01.1003
#> 3149      1962-01-01    1962-01-01 1962-01-01.1004
#> 3150      1962-01-01    1962-01-01 1962-01-01.1004
#> 3151      2006-08-31    1962-01-01 1962-01-01.1005
#> 3152      2006-08-31    1962-01-01 1962-01-01.1005
#> 3153      2000-12-31    1962-01-01 1962-01-01.1006
#> 3154      2000-12-31    1962-01-01 1962-01-01.1007
#> 3155      1962-01-01    1962-01-01 1962-01-01.1008
#> 3156      2000-12-31    1963-01-01 1963-01-01.1009
#> 3157      2000-12-31    1963-05-24 1963-05-24.1010
#> 3158      2000-12-31    1963-05-24 1963-05-24.1010
#> 3159      2008-12-31    1964-01-01 1964-01-01.1011
#> 3160      2008-12-31    1964-01-01 1964-01-01.1011
#> 3161      2002-12-31    1964-03-10 1964-03-10.1012
#> 3162      2010-12-31    1964-06-01 1964-06-01.1013
#> 3163      2010-12-31    1964-06-01 1964-06-01.1013